Correcciones en modelos y migraciones del backend. Actualizado el .env.example.

+++ Backend/app/Models/Vote.php
@@ -24,7 +24,7 @@ class Vote extends Model
     ];
 
     protected $casts = [
-        'score' => 'integer',
+        'score' => 'float',
     ];
 
     protected $hidden = [
